HIVE-23965: Improve plan regression tests using TPCDS30TB metastore dump and custom configs
1. Use Dockerized postgres metastore with TPC-DS 30TB dump
2. Use Hive config properties obtained and curated from real-life usages
3. Allow AbstractCliConfig to override metastore DB type
4. Rework CorePerfCliDriver to allow pre-initialized metastores

Extract TPCDS specific code to subclass and document appropriately
clarifying inaccuracies regarding the size of the tables in the javadoc.

Remove system property settings in the initialization of the driver and
leave in the configuration to set it up if needed. This is necessary to
be able to use the driver with a preinitialised metastore.

Remove redundant logs in System.err. Logging and throwing an exception
is an anti-pattern.

Replace assertions with exceptions and improve the messages.

5. Give more meaningful names to TPCDS related CLI configurations
6. Disable queries 30, 74, 84 with appropriate JIRA reference
7. Add TPC-DS query plans for the new driver (TestTezTPCDS30TBCliDriver)
8. Consider hive.current.database property when returning the session's
current db to avoid prefixing every query with the name of the database
9. Use maven build property to set postgres version
10. Upgrade postgres JDBC driver to version 42.2.14 to be compatible
with the docker image used.