HIVE-24031: Add test with big query reproducing the problem
The commit is merely for demonstrating the problem. It is not meant to
to be merged onto master since the query itself takes up a few MB and
also requires some changes in the default configurations.

1. Remove maskPatterns from QTestUtil#checkCliDriverResults since it is
VERY slow on big files.
2. Remove PostExecutePrinter since it triggers transformations for
computing lineage information that are expensive and not strictly
related to the problem.
3. Increase heap size to 8gb since otherwise there is an OOM.