Avoid accidental replace of standard aggs with Hive aggs
The final plan sometimes contains standard Calcite operators such as COUNT. When deserializing we don't want to change the operator since this alters the plan.

Ideally, we shouldn't have the same operator (eg., Calcite COUNT vs Hive count) in different variants in the plans but that's out of scope for the deserialization work.